Extract shared createProjectStatusUi into webview/common
Closes #354

All three platform webviews (simple, tec1, tec1g) now use the same
`createProjectStatusUi` abstraction from `webview/common/project-status-ui.ts`
instead of duplicating the setup-card/target-dropdown/project-root wiring inline.
`tec1g-project-status-ui.ts` becomes a thin re-export wrapper so existing
callers are unaffected.
