+++ webview/simple/index.ts
@@ -5,13 +5,10 @@
 import { appendSerialText } from '../common/serial';
 import { MemoryPanel } from '../common/memory-panel';
 import { applyInitializedProjectControls } from '../common/project-controls';
-import { resolveProjectViewState } from '../common/project-state';
 import { createSessionStatusController } from '../common/session-status';
 import { wireStopOnEntryControl } from '../common/stop-on-entry-control';
-import { createProjectRootButtonController } from '../common/project-root-button';
-import { resolveSetupCardState } from '../common/setup-card-state';
+import { createProjectStatusUi } from '../common/project-status-ui';
 import { acquireVscodeApi } from '../common/vscode';
-import { sendCreateProject } from '../common/create-project';
 import type { ProjectStatusPayload } from '../../src/contracts/platform-view';
 
 const TERMINAL_MAX = 8000;
@@ -45,17 +42,20 @@ const terminalOutEl = document.getElementById('terminalOut') as HTMLElement | nu
 const terminalClearEl = document.getElementById('terminalClear') as HTMLElement | null;
 
 let activeTab: 'ui' | 'memory' = 'ui';
-let currentRootPath = '';
 let projectIsInitialized = false;
-let currentRoots: Array<{ name: string; path: string; hasProject: boolean }> = [];
-let setupPrimaryActionType: 'openWorkspaceFolder' | 'selectProject' | 'createProject' =
-  'openWorkspaceFolder';
 let memoryRowSize = 16;
 let resizeTimer: number | null = null;
 
 const sessionStatusController = createSessionStatusController(vscode, restartDebugButton);
 const stopOnEntryControl = wireStopOnEntryControl(vscode, stopOnEntryInput);
-const projectRootController = createProjectRootButtonController(vscode, selectProjectButton);
+const projectStatusUi = createProjectStatusUi(vscode, {
+  selectProjectButton,
+  setupCard,
+  setupCardText,
+  setupPrimaryAction,
+  platformInitButton,
+  homeTargetSelect,
+}, 'simple');
 
 addWorkspaceFolderButton?.addEventListener('click', () => {
   vscode.postMessage({ type: 'openWorkspaceFolder' });
@@ -67,76 +67,12 @@ platformSelectEl?.addEventListener('change', () => {
   }
 });
 
-setupPrimaryAction?.addEventListener('click', () => {
-  if (setupPrimaryActionType === 'openWorkspaceFolder') {
-    vscode.postMessage({ type: 'openWorkspaceFolder' });
-    return;
-  }
-  if (setupPrimaryActionType === 'selectProject') {
-    vscode.postMessage({ type: 'selectProject' });
-    return;
-  }
-  sendCreateProject(vscode, 'simple');
-});
-
-platformInitButton?.addEventListener('click', () => {
-  sendCreateProject(vscode, 'simple');
-});
-
-homeTargetSelect?.addEventListener('change', () => {
-  const targetName = homeTargetSelect.value;
-  if (!targetName) {
-    return;
-  }
-  vscode.postMessage({ type: 'selectTarget', rootPath: currentRootPath, targetName });
-});
-
 terminalClearEl?.addEventListener('click', () => {
   if (terminalOutEl) {
     terminalOutEl.textContent = '';
   }
   vscode.postMessage({ type: 'serialClear' });
 });
-
-function clearSelectOptions(select: HTMLSelectElement): void {
-  while (select.firstChild) {
-    select.removeChild(select.firstChild);
-  }
-}
-
-function setSelectPlaceholder(select: HTMLSelectElement, label: string): void {
-  const option = document.createElement('option');
-  option.value = '';
-  option.textContent = label;
-  option.disabled = true;
-  option.selected = true;
-  select.appendChild(option);
-}
-
-function setTargetOptions(
-  options: Array<{ name: string; description?: string; detail?: string }>,
-  selectedTargetName?: string
-): void {
-  if (!homeTargetSelect) {
-    return;
-  }
-  clearSelectOptions(homeTargetSelect);
-  if (options.length === 0) {
-    setSelectPlaceholder(homeTargetSelect, 'No targets available');
-    homeTargetSelect.disabled = true;
-    return;
-  }
-  setSelectPlaceholder(homeTargetSelect, 'Select target...');
-  for (const option of options) {
-    const el = document.createElement('option');
-    el.value = option.name;
-    el.textContent = option.name;
-    el.title = option.detail ?? option.description ?? option.name;
-    homeTargetSelect.appendChild(el);
-  }
-  homeTargetSelect.disabled = false;
-  homeTargetSelect.value = selectedTargetName ?? '';
-}
 
 function applyProjectStatus(payload: {
   rootPath?: ProjectStatusPayload['rootPath'];
@@ -148,16 +84,7 @@ function applyProjectStatus(payload: {
   hasProject?: ProjectStatusPayload['hasProject'];
   stopOnEntry?: ProjectStatusPayload['stopOnEntry'];
 }): void {
-  const projectState = resolveProjectViewState(payload);
-  const initializedProject = projectState === 'initialized';
-  currentRootPath = payload.rootPath ?? '';
-  currentRoots = payload.roots ?? [];
-  projectRootController.applyProjectStatus({
-    rootPath: payload.rootPath,
-    roots: payload.roots ?? [],
-    targetCount: payload.targets?.length ?? 0,
-  });
-  setTargetOptions(initializedProject ? (payload.targets ?? []) : [], payload.targetName);
+  projectStatusUi.applyProjectStatus(payload);
   if (platformSelectEl && payload.platform !== undefined) {
     platformSelectEl.value = payload.platform;
   }
@@ -181,29 +108,6 @@ function applyProjectStatus(payload: {
     hasProject: initialized,
     stopOnEntry: payload.stopOnEntry,
   });
-  const selected = currentRoots.find((r) => r.path === currentRootPath) ?? currentRoots[0];
-  const targetCount = payload.targets?.length ?? 0;
-  if (!setupCard || !setupCardText || !setupPrimaryAction) {
-    return;
-  }
-  const setupState = resolveSetupCardState(
-    selected,
-    projectState,
-    targetCount,
-    currentRoots.length
-  );
-  if (setupState === null) {
-    setupCard.hidden = true;
-    return;
-  }
-  setupCard.hidden = false;
-  setupPrimaryActionType = setupState.primaryAction;
-  setupCardText.textContent = setupState.text;
-  // When createProject is the pending action, platformInitButton is the primary CTA;
-  // hide the setup card button to avoid showing two equivalent "Initialize" actions.
-  const isCreateProject = setupState.primaryAction === 'createProject';
-  setupPrimaryAction.hidden = isCreateProject;
-  setupPrimaryAction.textContent = isCreateProject ? '' : setupState.primaryLabel;
 }
 
 applyProjectStatus({});
@@ -338,5 +242,5 @@ window.addEventListener('resize', () => scheduleMemoryResize());
 window.addEventListener('beforeunload', () => {
   sessionStatusController.dispose();
   stopOnEntryControl.dispose();
-  projectRootController.dispose();
+  projectStatusUi.dispose();
 });
